Rename table::* -> table::Loaded*
To avoid name clashes with the structs in chunks.rs, prefix the structs
in table.rs with 'Loaded'. The name indicates that these structs are
higher level than the raw chunks (and libandroidfw uses the same
convention!).

On a module level, export LoadedTable as just Table: chunks are hidden
from the caller so no need to introduce states in the public names.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>