Implement ResourceId::from_u32
Add a conversion from u32 to ResourceId. Limit its visibility to within
the arsc crate: it is handy for unit tests and when converting the raw
arsc data but should not be visible to users of the crate.

Also, convert some of the unit tests to use ResourceId::from_u32 instead
of ResourceId::from_parts to increase readability.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>